+++ spine-ios/Example/Spine iOS Example/spine-ios/SpineViewController.swift
@@ -21,16 +21,16 @@ public final class SpineViewController: UIViewController {
     private let skeletonFile: String
     private let controller: SpineController
     
-    private let contentMode: SpineContentMode
-    private let alignment: SpineAlignment
+    private let contentMode: Spine.ContentMode
+    private let alignment: Spine.Alignment
     private let boundsProvider: BoundsProvider
     
     public init(
         atlasFile: String,
         skeletonFile: String,
         controller: SpineController,
-        contentMode: SpineContentMode? = nil,
-        alignment: SpineAlignment? = nil,
+        contentMode: Spine.ContentMode? = nil,
+        alignment: Spine.Alignment? = nil,
         boundsProvider: BoundsProvider? = nil
     ) {
         self.atlasFile = atlasFile
