+++ spine-ios/Sources/Spine/BoundsProvider.swift
@@ -1,0 +1,135 @@
+import Foundation
+import CoreGraphics
+
+/// Base class for bounds providers. A bounds provider calculates the axis aligned bounding box
+/// used to scale and fit a skeleton inside the bounds of a [SpineWidget].
+public protocol BoundsProvider {
+    func computeBounds(for drawable: SkeletonDrawableWrapper) -> CGRect
+}
+
+/// A [BoundsProvider] that calculates the bounding box of the skeleton based on the visible
+/// attachments in the setup pose.
+public final class SetupPoseBounds: BoundsProvider {
+    
+    public init() {}
+
+    public func computeBounds(for drawable: SkeletonDrawableWrapper) -> CGRect {
+        return CGRect(bounds: drawable.skeleton.bounds)
+    }
+}
+
+/// A [BoundsProvider] that calculates the bounding box needed for a combination of skins
+/// and an animation.
+public final class SkinAndAnimationBounds: BoundsProvider {
+    
+    private let animation: String?
+    private let skins: [String]
+    private let stepTime: TimeInterval;
+
+    /// Constructs a new provider that will use the given [skins] and [animation] to calculate
+    /// the bounding box of the skeleton. If no skins are given, the default skin is used.
+    /// The [stepTime], given in seconds, defines at what interval the bounds should be sampled
+    /// across the entire animation.
+    public init(animation: String, skins: [String]? = nil, let stepTime: TimeInterval = 0.1) {
+        self.animation = animation
+        if let skins, !skins.isEmpty {
+            self.skins = skins
+        } else {
+            self.skins = ["default"]
+        }
+        self.stepTime = stepTime
+    }
+    
+    public func computeBounds(for drawable: SkeletonDrawableWrapper) -> CGRect {
+        let data = drawable.skeleton.data
+        let oldSkin: Skin? = drawable.skeleton.skin
+        let customSkin = Skin.create(name: "custom-skin")
+        for skinName in skins {
+            let skin = data.findSkin(name: skinName) // TODO: Introduce optionsals where needed
+            if let skin = data.findSkin(name: skinName) {
+                customSkin.addSkin(other: skin)
+            }
+        }
+        drawable.skeleton.skin = customSkin
+        drawable.skeleton.setToSetupPose();
+
+        let animation = animation.flatMap { data.findAnimation(name: $0) }
+        var minX = Float.Magnitude.greatestFiniteMagnitude
+        var minY = Float.Magnitude.greatestFiniteMagnitude
+        var maxX = -Float.Magnitude.greatestFiniteMagnitude
+        var maxY = -Float.Magnitude.greatestFiniteMagnitude
+        if let animation {
+            drawable.animationState.setAnimation(trackIndex: 0, animation: animation, loop: false)
+            let steps = Int(max(Double(animation.duration) / stepTime, 1.0))
+            for i in 0..<steps {
+                drawable.update(delta: i > 0 ? Float(stepTime) : 0.0)
+                let bounds = drawable.skeleton.bounds;
+                minX = min(minX, bounds.x)
+                minY = min(minY, bounds.y)
+                maxX = max(maxX, minX + bounds.width)
+                maxY = max(maxY, minY + bounds.height)
+            }
+        } else {
+            let bounds = drawable.skeleton.bounds;
+            minX = bounds.x
+            minY = bounds.y
+            maxX = minX + bounds.width
+            maxY = minY + bounds.height
+        }
+        drawable.skeleton.setSkinByName(skinName: "default");
+        drawable.animationState.clearTracks();
+        
+        if let oldSkin {
+            drawable.skeleton.skin = oldSkin
+        }
+        drawable.skeleton.setToSetupPose();
+        drawable.update(delta: 0);
+        customSkin.dispose();
+        return CGRectMake(CGFloat(minX), CGFloat(minY), CGFloat(maxX - minX), CGFloat(maxY - minY));
+      }
+}
+
+public enum ContentMode {
+    case fit
+    case fill
+}
+
+public enum Alignment {
+    case topLeft
+    case topCenter
+    case topRight
+    case centerLeft
+    case center
+    case centerRight
+    case bottomLeft
+    case bottomCenter
+    case bottomRight
+    
+    public var x: CGFloat {
+        switch self {
+        case .topLeft: return -1.0
+        case .topCenter: return 0.0
+        case .topRight: return 1.0
+        case .centerLeft: return -1.0
+        case .center: return 0.0
+        case .centerRight: return 1.0
+        case .bottomLeft: return -1.0
+        case .bottomCenter: return 0.0
+        case .bottomRight: return 1.0
+        }
+    }
+    
+    public var y: CGFloat {
+        switch self {
+        case .topLeft: return -1.0
+        case .topCenter: return -1.0
+        case .topRight: return -1.0
+        case .centerLeft: return -0.0
+        case .center: return 0.0
+        case .centerRight: return 0.0
+        case .bottomLeft: return -1.0
+        case .bottomCenter: return 1.0
+        case .bottomRight: return 1.0
+        }
+    }
+}
